Fix authApi import warning
