+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Split the browser-session `src/services/authApi.ts` login/bootstrap client from the statically imported account/MFA helpers so Vite/Rolldown no longer emits the `INEFFECTIVE_DYNAMIC_IMPORT` warning for `authApi` during `npm run build`. Shared auth API error/JSON parsing now lives in `src/services/authApiShared.ts`, account-scoped calls live in `src/services/authAccountApi.ts`, and the login/bootstrap lazy-import boundary remains effective.
 - Reworked the public browser-session bootstrap path so `/login` no longer issues an unauthenticated `/v1/me` probe when there is no local auth snapshot and no `XSRF-TOKEN` cookie, eliminating the spurious 401 console error on logged-out refreshes while keeping protected-route recovery and authenticated session restoration intact.
 - Split the public login flow from the authenticated shell so logged-out loads no longer flash post-login navigation skeletons, heavy private route chunks, or browser passkey capability rows after first paint. The login route now uses dedicated public loading states, lazy-loads MFA/passkey/auth modules where safe, and keeps the visible card stable during session completion to remove mobile reload jitter.
 - Restored the visible login logo while preserving the Lighthouse image fixes by rendering higher-resolution raster assets with their real intrinsic aspect ratios instead of stretching the small square PNG placeholders.
