+++ src/services/authApi.ts
@@ -4,23 +4,14 @@
 import type {
   CompletedLoginResponse,
   LoginMfaChallengeResponse,
-  MfaRecoveryCodeRevealResponse,
-  MfaStatusResponse,
-  MfaTotpEnrollmentResponse,
   MfaVerificationCodeRequest,
-  PasskeyDeletionResponse,
   PasskeyAuthenticationChallengeResponse,
   PasskeyAuthenticationVerificationRequest,
-  PasskeyListResponse,
-  PasskeyRegistrationChallengeResponse,
-  PasskeyRegistrationResponse,
-  PasskeyRegistrationVerificationRequest,
   SessionLoginResponse,
-  TotpCodeRequest,
-  VerificationNotificationResponse,
 } from "@/types/api";
 import { ApiBaseUrlConfigurationError, buildApiUrl } from "../config";
 import { AuthApiError } from "./AuthApiError";
+import { createAuthApiError, parseJsonResponse } from "./authApiShared";
 import { fetchCsrfToken, apiFetch } from "./csrf";
 
 interface LoginCredentials {
@@ -33,102 +24,19 @@ export type BrowserLoginResponse =
   | SessionLoginResponse
   | LoginMfaChallengeResponse;
 
-interface ApiError {
-  message: string;
-  errors?: Record<string, string[]>;
-  code?: string;
-}
-
-function hasJsonContentType(response: Response): boolean {
-  if (
-    !("headers" in response) ||
-    !response.headers ||
-    typeof response.headers.get !== "function"
-  ) {
-    return typeof response.json === "function";
-  }
-
-  const contentType = response.headers.get("Content-Type");
-
-  return contentType?.toLowerCase().includes("application/json") ?? false;
-}
-
-async function parseJsonResponse<T>(
-  response: Response,
-  operation: string
-): Promise<T> {
-  if (!hasJsonContentType(response)) {
-    throw new AuthApiError(
-      `${operation}: expected application/json response from API`
-    );
-  }
-
-  try {
-    return (await response.json()) as T;
-  } catch {
-    throw new AuthApiError(`${operation}: received malformed JSON from API`);
-  }
-}
-
-async function parseJsonError(response: Response): Promise<ApiError | null> {
-  if (!hasJsonContentType(response)) {
-    return null;
-  }
-
-  try {
-    return (await response.json()) as ApiError;
-  } catch {
-    return null;
-  }
-}
-
-function parseRetryAfterSeconds(response: Response): number | undefined {
-  if (
-    !("headers" in response) ||
-    !response.headers ||
-    typeof response.headers.get !== "function"
-  ) {
-    return undefined;
-  }
-
-  const retryAfterHeader = response.headers.get("Retry-After");
-  const retryAfterSeconds = retryAfterHeader
-    ? Number.parseInt(retryAfterHeader, 10)
-    : Number.NaN;
-
-  return Number.isFinite(retryAfterSeconds) && retryAfterSeconds >= 0
-    ? retryAfterSeconds
-    : undefined;
-}
-
 export { AuthApiError } from "./AuthApiError";
-
-async function createAuthApiError(
-  response: Response,
-  defaultMessage: string,
-  nonJsonMessage = `${defaultMessage}: ${response.status} ${response.statusText}`
-): Promise<AuthApiError> {
-  const error = await parseJsonError(response);
-  const retryAfterSeconds = parseRetryAfterSeconds(response);
-
-  if (!error) {
-    return new AuthApiError(
-      nonJsonMessage,
-      undefined,
-      response.status,
-      undefined,
-      retryAfterSeconds
-    );
-  }
-
-  return new AuthApiError(
-    error.message || defaultMessage,
-    error.errors,
-    response.status,
-    error.code,
-    retryAfterSeconds
-  );
-}
+export {
+  confirmTotpEnrollment,
+  deletePasskey,
+  disableMfa,
+  getMfaStatus,
+  getPasskeys,
+  regenerateRecoveryCodes,
+  sendVerificationNotification,
+  startPasskeyRegistrationChallenge,
+  startTotpEnrollment,
+  verifyPasskeyRegistrationChallenge,
+} from "./authAccountApi";
 
 /**
  * Login with email and password
@@ -314,248 +222,4 @@ export async function getCurrentUser(): Promise<SessionLoginResponse["user"]> {
     response,
     "Current user fetch failed"
   );
-}
-
-export async function sendVerificationNotification(): Promise<VerificationNotificationResponse> {
-  await fetchCsrfToken();
-
-  const response = await apiFetch(
-    buildApiUrl("/v1/auth/email/verification-notification"),
-    {
-      method: "POST",
-      cache: "no-store",
-      headers: {
-        Accept: "application/json",
-      },
-    }
-  );
-
-  if (!response.ok) {
-    throw await createAuthApiError(
-      response,
-      "Verification email request failed"
-    );
-  }
-
-  return parseJsonResponse<VerificationNotificationResponse>(
-    response,
-    "Verification email request failed"
-  );
-}
-
-export async function getMfaStatus(): Promise<MfaStatusResponse> {
-  const response = await apiFetch(buildApiUrl("/v1/me/mfa"), {
-    method: "GET",
-    cache: "no-store",
-    headers: {
-      Accept: "application/json",
-    },
-  });
-
-  if (!response.ok) {
-    throw await createAuthApiError(response, "MFA status fetch failed");
-  }
-
-  return parseJsonResponse<MfaStatusResponse>(
-    response,
-    "MFA status fetch failed"
-  );
-}
-
-export async function getPasskeys(): Promise<PasskeyListResponse> {
-  const response = await apiFetch(buildApiUrl("/v1/me/passkeys"), {
-    method: "GET",
-    cache: "no-store",
-    headers: {
-      Accept: "application/json",
-    },
-  });
-
-  if (!response.ok) {
-    throw await createAuthApiError(response, "Passkey list fetch failed");
-  }
-
-  return parseJsonResponse<PasskeyListResponse>(
-    response,
-    "Passkey list fetch failed"
-  );
-}
-
-export async function startPasskeyRegistrationChallenge(): Promise<PasskeyRegistrationChallengeResponse> {
-  await fetchCsrfToken();
-
-  const response = await apiFetch(
-    buildApiUrl("/v1/me/passkeys/challenges/registration"),
-    {
-      method: "POST",
-      cache: "no-store",
-      headers: {
-        Accept: "application/json",
-      },
-    }
-  );
-
-  if (!response.ok) {
-    throw await createAuthApiError(
-      response,
-      "Passkey registration challenge start failed"
-    );
-  }
-
-  return parseJsonResponse<PasskeyRegistrationChallengeResponse>(
-    response,
-    "Passkey registration challenge start failed"
-  );
-}
-
-export async function verifyPasskeyRegistrationChallenge(
-  challengeId: string,
-  payload: PasskeyRegistrationVerificationRequest
-): Promise<PasskeyRegistrationResponse> {
-  await fetchCsrfToken();
-
-  const response = await apiFetch(
-    buildApiUrl(
-      `/v1/me/passkeys/challenges/registration/${challengeId}/verify`
-    ),
-    {
-      method: "POST",
-      cache: "no-store",
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
-      },
-      body: JSON.stringify(payload),
-    }
-  );
-
-  if (!response.ok) {
-    throw await createAuthApiError(response, "Passkey registration failed");
-  }
-
-  return parseJsonResponse<PasskeyRegistrationResponse>(
-    response,
-    "Passkey registration failed"
-  );
-}
-
-export async function deletePasskey(
-  credentialId: string
-): Promise<PasskeyDeletionResponse> {
-  const response = await apiFetch(
-    buildApiUrl(`/v1/me/passkeys/${credentialId}`),
-    {
-      method: "DELETE",
-      cache: "no-store",
-      headers: {
-        Accept: "application/json",
-      },
-    }
-  );
-
-  if (!response.ok) {
-    throw await createAuthApiError(response, "Passkey deletion failed");
-  }
-
-  return parseJsonResponse<PasskeyDeletionResponse>(
-    response,
-    "Passkey deletion failed"
-  );
-}
-
-export async function startTotpEnrollment(): Promise<MfaTotpEnrollmentResponse> {
-  const response = await apiFetch(buildApiUrl("/v1/me/mfa/totp/enrollment"), {
-    method: "POST",
-    cache: "no-store",
-    headers: {
-      Accept: "application/json",
-    },
-  });
-
-  if (!response.ok) {
-    throw await createAuthApiError(response, "TOTP enrollment start failed");
-  }
-
-  return parseJsonResponse<MfaTotpEnrollmentResponse>(
-    response,
-    "TOTP enrollment start failed"
-  );
-}
-
-export async function confirmTotpEnrollment(
-  payload: TotpCodeRequest
-): Promise<MfaRecoveryCodeRevealResponse> {
-  const response = await apiFetch(
-    buildApiUrl("/v1/me/mfa/totp/enrollment/confirm"),
-    {
-      method: "POST",
-      cache: "no-store",
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
-      },
-      body: JSON.stringify(payload),
-    }
-  );
-
-  if (!response.ok) {
-    throw await createAuthApiError(
-      response,
-      "TOTP enrollment confirmation failed"
-    );
-  }
-
-  return parseJsonResponse<MfaRecoveryCodeRevealResponse>(
-    response,
-    "TOTP enrollment confirmation failed"
-  );
-}
-
-export async function regenerateRecoveryCodes(
-  payload: MfaVerificationCodeRequest
-): Promise<MfaRecoveryCodeRevealResponse> {
-  const response = await apiFetch(
-    buildApiUrl("/v1/me/mfa/recovery-codes/regenerate"),
-    {
-      method: "POST",
-      cache: "no-store",
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
-      },
-      body: JSON.stringify(payload),
-    }
-  );
-
-  if (!response.ok) {
-    throw await createAuthApiError(
-      response,
-      "Recovery code regeneration failed"
-    );
-  }
-
-  return parseJsonResponse<MfaRecoveryCodeRevealResponse>(
-    response,
-    "Recovery code regeneration failed"
-  );
-}
-
-export async function disableMfa(
-  payload: MfaVerificationCodeRequest
-): Promise<MfaStatusResponse> {
-  const response = await apiFetch(buildApiUrl("/v1/me/mfa"), {
-    method: "DELETE",
-    cache: "no-store",
-    headers: {
-      "Content-Type": "application/json",
-      Accept: "application/json",
-    },
-    body: JSON.stringify(payload),
-  });
-
-  if (!response.ok) {
-    throw await createAuthApiError(response, "MFA disable failed");
-  }
-
-  return parseJsonResponse<MfaStatusResponse>(response, "MFA disable failed");
 }
